Camera Rotation
Added limited camera rotation. How it works currently is that you click and hold, and while the mouse button is held the camera rotates based on how far your cursor is from the center. It also locks rotation round the x axis at Pi/2 radians so you can't do a loop-de-loop with it.

When I get more time to work on it I will make sure that the cursor is hidden and it can be moved without having to click. Additionally, rotation currently happens in the local coordinate system which needs to be fixed.